desloppify: post-review cleanup (5 clusters)
- **telemetry-vendor-neutral-residue**: move format_tokens,
  format_context_pct, and context_window_for_model/format tests
  from agents/claude/telemetry.rs into core::telemetry. Split the
  lib.rs re-export so vendor-neutral names come from core::telemetry
  and Claude-specific names stay on agents::claude::telemetry.
- **sibling-asymmetric-cleanup**: delete unused is_claude_auto_type;
  reorder lib.rs pub use blocks alphabetically; rename
  OpenCodeMessageTail::new -> for_session(&Path, _) matching sibling
  tails; expose pub snapshot field on OpenCodeMessageTail mirrored
  from watch.snapshot in poll() so the AppShell dispatch site reads
  uniformly across the four agents.
- **doc-drift-postcutover**: refresh docs/ARCHITECTURE.md core
  modules table; sweep stale claude_telemetry::/claude_discovery::
  cross-references in doc comments to agents::<vendor>::<kind>::;
  drop the nonexistent 'app crate' reference from the lib.rs header.
- **single-edit-cleanups**: switch new_worktree::submit to
  format!("{err:#}") so the anyhow chain reaches the dialog error
  slot; delete the unused theme::status_running superseded shim;
  merge the two near-identical folder pickers in new_project into a
  shared prompt_dialog_folder helper.
- **clippy-letchain-sweep**: cargo clippy --fix workspace-wide for
  collapsible_if; let-chain rewrites in src/app.rs,
  terminal/src/view.rs, terminal/src/paint.rs, build.rs.
- Bonus: distinguish "git missing" (Err) from "not a worktree" /
  "cwd doesn't exist" (Ok(None)) in git_status by going around
  run_git's wrap-non-zero-as-Err behaviour for the inside-worktree
  probe; updates the two non-repo tests to expect Ok(None).

All 508 workspace tests pass.

Resolves desloppify review findings shared_telemetry_reexported_via_claude,
vendor_neutral_formatters_in_claude_telemetry, claude_telemetry_reexport_residue,
core_telemetry_tests_left_behind_after_refactor,
stale_is_claude_auto_type_predicate, lib_rs_pub_use_alphabetisation_break,
opencode_tail_constructor_outlier, opencode_tail_snapshot_accessor_outlier,
architecture_doc_module_drift, stale_flat_agent_module_refs_in_doc_comments,
core_lib_doc_references_nonexistent_app_crate,
new_worktree_err_to_string_drops_chain, theme_status_running_superseded_shim,
near_duplicate_folder_pickers, and remaining_collapsible_if_cluster.
