+++ core/src/lib.rs
@@ -10,8 +10,8 @@
 //!   only have to change it here, not in two places.
 //!
 //! No gpui, alacritty, or windows-rs imports here. Anything UI-bound
-//! lives in the `app` crate. Anything terminal-protocol-bound lives in
-//! `codescope-terminal`.
+//! lives in the root `codescope-rs` binary crate (`src/`). Anything
+//! terminal-protocol-bound lives in `codescope-terminal`.
 
 pub mod agent;
 pub mod agent_registry;
@@ -44,10 +44,7 @@ pub use agent_registry::{
     built_in_defaults,
 };
 pub use agents::claude::discovery::POLL_INTERVAL_MS as CLAUDE_DISCOVERY_POLL_MS;
-pub use agents::claude::telemetry::{
-    ClaudeTranscriptTail, SessionState, TelemetrySnapshot, context_window_for_model, encode_cwd,
-    format_context_pct, format_tokens, model_display_name,
-};
+pub use agents::claude::telemetry::{ClaudeTranscriptTail, encode_cwd, model_display_name};
 pub use agents::copilot::discovery::POLL_INTERVAL_MS as COPILOT_DISCOVERY_POLL_MS;
 pub use agents::copilot::telemetry::CopilotTranscriptTail;
 pub use agents::opencode::discovery::POLL_INTERVAL_MS as OPENCODE_DISCOVERY_POLL_MS;
@@ -55,17 +52,20 @@ pub use agents::opencode::telemetry::OpenCodeMessageTail;
 pub use agents::pi::discovery::POLL_INTERVAL_MS as PI_DISCOVERY_POLL_MS;
 pub use agents::pi::telemetry::PiTranscriptTail;
 pub use attachments::{SavedAttachment, save_attachment_bytes};
+pub use layout::{LayoutState, RestoreTab, SessionPlacement};
 pub use overview::{
     LiveSessionLookup, OverviewLifecycle, OverviewRow,
     build_rows as build_overview_rows,
     build_rows_for_live as build_overview_rows_for_live,
 };
-pub use layout::{LayoutState, RestoreTab, SessionPlacement};
 pub use paths::AppPaths;
 pub use projects::{Project, ProjectsConfig, Session, Worktree};
 pub use session::{RetentionPolicy, SessionDescriptor, SessionManager, build_agent_shell_args, now_iso8601};
 pub use settings::{CursorSettings, CursorShape, DEFAULT_AGENT_ID, FontSettings, Settings};
 pub use tab_drag::{TabRect, compute_drop_index};
 pub use tab_title::{TAB_TITLE_SEPARATOR, rebuild_title};
+pub use telemetry::{
+    SessionState, TelemetrySnapshot, context_window_for_model, format_context_pct, format_tokens,
+};
 pub use theme::{Rgb, Theme, ThemePalette, builtin};
 pub use window_state::WindowState;
